Fix bug caused by circular dependency of PeerAction enum
