+++ src/core/peer/PeerConnectionManager.ts
@@ -81,8 +81,6 @@ export default class PeerConnectionManager {
     );
 
     if (!res.success) {
-      console.log(`[DEBUG] Response:`);
-      console.log(res);
       const { name, message } = res.error;
       throw new GeneralError(message, name);
     }
