+++ src/core/peer/types/PeerAction.ts
@@ -1,0 +1,9 @@
+export enum PeerAction {
+  PING = "PING",
+  JOIN = "JOIN",
+  RECONNECT = "RECONNECT",
+  PUSH_SHARED = "PUSH_SHARED",
+  PULL_SHARED = "PULL_SHARED",
+  PUSH_PRIVATE = "PUSH_PRIVATE",
+  END_GAME = "END_GAME",
+}
